Create Account
Finished the create account section for patients. There is checks in place to make sure the form is filled out correctly. If the form is filled out properly it will check with firebase to see if the email has been used and if not will create a new user and the website will redirect the user to the account profile page.